Improve tests using rstest, move into own files, add more tests
